fix: teminatOrani int yap (Evrim Integer bekliyor)
Evrim API "Invalid class typecast for property teminatOrani
[Expected: Integer, Actual: Float]" hatasi veriyordu.

+++ src/Orko.Portal.Infrastructure/ExternalServices/EvrimModels/EvrimDeclarationRequest.cs
@@ -232,7 +232,7 @@ public class EvrimDeclarationRequest
     public string? TeminatSekli { get; set; }
 
     [JsonPropertyName("teminatOrani")]
-    public decimal? TeminatOrani { get; set; }
+    public int? TeminatOrani { get; set; }  // Evrim Integer bekler
 
     [JsonPropertyName("teminatAciklama")]
     public string? TeminatAciklama { get; set; }
